requirements.txt
Flask>=2.2,<3.0
moviepy>=1.0.3,<2.0
Pillow>=9.0
imageio-ffmpeg>=0.4.7
numpy>=1.23
tqdm>=4.64
python-dotenv>=0.21
gunicorn>=20.1
